refactor: split lifecycle_builder.py into builder + helpers
