+++ tests/unit/api/test_auto_promote.py
@@ -7,7 +7,7 @@ import pytest
 from synthorg.api.approval_store import ApprovalStore
 from synthorg.api.auth.models import OrgRole, User
 from synthorg.api.guards import HumanRole
-from synthorg.api.lifecycle_builder import _maybe_promote_first_owner
+from synthorg.api.lifecycle_helpers import _maybe_promote_first_owner
 from synthorg.api.state import AppState
 from synthorg.config.schema import RootConfig
 from tests.unit.api.fakes import FakePersistenceBackend
